my-cdk-app/requirements.txt
aws-cdk-lib==2.172.0
constructs>=10.0.0,<11.0.0
torch==2.4.1
torchvision==0.19.1
hydra-colorlog==1.2.0
hydra-core==1.3.2
lightning[extra]==2.4.0
loguru==0.7.2
rich==13.8.1
rootutils==1.0.7
tensorboard==2.17.1
timm==1.0.9
pandas>=2.2.3
numpy>=1.26.0
transformers>=4.45.2
aim>=3.25.0
mlflow>=2.17.1
hydra-optuna-sweeper>=1.2.0
aiologger>=0.7.0
pyyaml>=6.0.2
dvc-s3>=3.2.0
litserve>=0.2.4
gpustat>=1.1.1
nvitop>=1.3.2
gradio==5.7.1
gradio-client>=1.5.0
accelerate>=1.1.1
cryptography>=44.0.0
boto3
pyopenssl>=24.3.0
